+++ lib/catan/entities/player.rb
@@ -4,13 +4,14 @@ class Player
   EmptyName = Class.new(ArgumentError)
   InvalidColor = Class.new(ArgumentError)
 
-  attr_reader :index, :name, :color, :resources
+  attr_reader :index, :name, :color, :resources, :cards
 
   def initialize(index:, name:, color:, resources: Resources.create_empty_bank)
     @index = index
     @name = name
     @color = color.to_sym
     @resources = resources
+    @cards = []
     raise EmptyName unless name_valid?
     raise InvalidColor unless color_valid?
   end
@@ -21,6 +22,14 @@ class Player
 
   def color_valid?
     %i[orange red white blue].include?(@color)
+  end
+
+  def pay(cost)
+    @resources.substitute(:brick, cost.brick)
+    @resources.substitute(:lumber, cost.lumber)
+    @resources.substitute(:wool, cost.wool)
+    @resources.substitute(:grain, cost.grain)
+    @resources.substitute(:ore, cost.ore)
   end
 
   def ==(other)