+++ spec/spec_helper.rb
@@ -5,6 +5,7 @@ require_relative File.join('.', 'shared', 'interactions.rb')
 require_relative File.join('.', 'test_interactors', 'set_turn.rb')
 require_relative File.join('.', 'test_interactors', 'take_action.rb')
 require_relative File.join('.', 'test_interactors', 'build_settlement.rb')
+require_relative File.join('.', 'test_interactors', 'set_resources.rb')
 
 RSpec.configure do |config|
   config.before(:example) do